feat(qwik-nx): qwikNxVite plugin to include only project's dependencies  by default
