+++ packages/qwik-nx/src/plugins/index.ts
@@ -1,1 +1,5 @@
 export * from './qwik-nx-vite.plugin';
+export type {
+  ProjectFilter,
+  QwikNxVitePluginOptions,
+} from './models/qwik-nx-vite';
